dom/src/threads/test/test_recursion.html

<!DOCTYPE HTML>
<html>
<!--
Tests of DOM Worker Threads
-->
<head>
  <title>Test for DOM Worker Threads Recursion</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

  var worker = new Worker("recursion_worker.js");

  worker.onerror = function(event) {
    is(event.data,
       '[JavaScript Error: "too much recursion" {file: "http://localhost:8888' +
       '/tests/dom/src/threads/tests/recursion_worker.js" line: 2}]');
    SimpleTest.finish();
  }

  worker.postMessage("start");

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
